volatile_memory: Add helpers to copy to/from volatile slice
Add separate helpers to copy from and to the volatile slice.

This allows the real logic to be present at a single place, which can be
easily updated for special cases later on. For example extra memory
mapping for systems where the regions can't be mapped in advance, like
in case of Xen's Grant mappings.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>