feat: Introduce Read/WriteVolatile traits
With #217, various `read_from` and `write_to` methods across vm-memory
incurred a performance penalty due to an unneccesary copy of all data
read and written. This is due to these functions operating on arbitrary
Read/Write implementations, which cannot generally be assumed to respect
the volatile semantics of guest memory. Additionally, they required
taking slices to guest memory, which was undefined behavior.

This patch series fixes those short-comings by introducing a pair of new
traits, `ReadVolatile` and `WriteVolatile` which are identical to the
standard library's `Read` and `Write` traits, with the difference being
that they operate on `VolatileSlice`s instead of `&mut [u8]`. Using
these traits will therefore not incur the performance penalty of #217.
Additionally, this will make using the vm-memory library easier for
users familiar with std::io, as the APIs will be more familiar (in
particular, the unintuitive "swapping" of reader/writer src and dst of
the old read_from/write_to APIs is resolved).

Please note that `impl<T: AsRawFd> ReadVolatile for T` is not be
possible, as the orphan rules will note that "upstream crates may add a
new impl of trait std::os::fd::AsRawFd for type &[u8] in future
versions".

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>